bugfix: bad start/end address of cmd symbolic table
after upgrading my system, the start address and the end address of
the command symbolic table became upside down. there must be sth
wrong with the link script!!

helper.mk -> include/u-boot.lst -> u-boot.lds

the problems is, in helper.mk, the output sequence of the chars '\a',
'@' and '~' are different in different OSes when processed by 'sort'.
then use them to generate related script brings troubles.

to fix the problem, i replace them to '#1', '#2', '#3'.